Configurable http.Client timeout for tokenHandler
The registry/client/auth package contains an AuthenticationHandler
interface which is meant to modify a registry-bound HTTP request
with authentication information.

The tokenHandler implementation of this interface may make outbound
HTTP requests in order to fetch the token to be populated into the
original request. These auxiliary requests are made using ephemeral
http.Client instances which are created via the (*tokenHandler).client()
function. At present, the http.Clients created by this function have
a hard-coded timeout of 15 seconds.

A larger piece of work calls for the timeout associated with certain
outbound HTTP requests by dockerd to be made configurable. This commit
makes the timeout of the above-mentioned ephemeral http.Client instances
configurable at construction via a field in the TokenHandlerOptions
struct. In order to retain backwards compatibility with any existing
library consumers, the field is passed as a pointer within the options
struct, and nil values are handled in the 'WithOptions' constructor. If
the default constructor is called, or if the 'HttpTimeout' field of
the TokenHandlerOptions struct is nil, then a default http timeout of
15 seconds is used.